+++ src/components/probe.rs
@@ -1,6 +1,7 @@
 use crate::common::{Component, Id, Input, OutputType, Ports};
 use log::*;
 use serde::{Deserialize, Serialize};
+use std::rc::Rc;
 #[derive(Serialize, Deserialize)]
 pub struct Probe {
     pub id: Id,
@@ -25,5 +26,19 @@ impl Component for Probe {
                 vec![],
             ),
         )
+    }
+}
+
+impl Probe {
+    pub fn new(id: &str, pos: (f32, f32), input: Input) -> Self {
+        Probe {
+            id: id.to_string(),
+            pos,
+            input,
+        }
+    }
+
+    pub fn rc_new(id: &str, pos: (f32, f32), input: Input) -> Rc<Self> {
+        Rc::new(Probe::new(id, pos, input))
     }
 }