+++ src/components/probe_stim.rs
@@ -29,9 +29,11 @@ impl Component for ProbeStim {
     }
 
     fn clock(&self, simulator: &mut Simulator) {
-        // trace!("-- clock --", simulator.clock)
-        simulator.set_out_val(&self.id, "out", self.values[0]);
+        trace!("-- cycle {} --", simulator.cycle);
+        simulator.set_out_val(&self.id, "out", self.values[simulator.cycle]);
     }
+
+    // notice we don't implement `un_clock` since the state is already kept in history
 }
 
 impl ProbeStim {
@@ -45,5 +47,76 @@ impl ProbeStim {
 
     pub fn rc_new(id: &str, pos: (f32, f32), values: Vec<impl Into<Signal>>) -> Rc<Self> {
         Rc::new(ProbeStim::new(id, pos, values))
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use crate::common::{ComponentStore, Input};
+
+    #[test]
+    fn test_probe_stim() {
+        let cs = ComponentStore {
+            store: vec![ProbeStim::rc_new("stim", (0.0, 0.0), vec![0, 1, 2, 3])],
+        };
+
+        let mut simulator = Simulator::new(&cs);
+        // output
+        let out = &Input::new("stim", "out");
+
+        // reset
+        println!("<reset>");
+        println!("sim_state {:?}", simulator.sim_state);
+        assert_eq!(simulator.cycle, 1);
+        assert_eq!(simulator.get_input_val(out), 0.into());
+
+        println!("<clock>");
+        simulator.clock();
+        println!("sim_state {:?}", simulator.sim_state);
+        assert_eq!(simulator.cycle, 2);
+        assert_eq!(simulator.get_input_val(out), 1.into());
+
+        println!("<clock>");
+        simulator.clock();
+        println!("sim_state {:?}", simulator.sim_state);
+        assert_eq!(simulator.cycle, 3);
+        assert_eq!(simulator.get_input_val(out), 2.into());
+
+        println!("<un_clock>");
+        simulator.un_clock();
+        println!("sim_state {:?}", simulator.sim_state);
+        assert_eq!(simulator.cycle, 2);
+        assert_eq!(simulator.get_input_val(out), 1.into());
+
+        println!("<un_clock>");
+        simulator.un_clock();
+        println!("sim_state {:?}", simulator.sim_state);
+        assert_eq!(simulator.cycle, 1);
+        assert_eq!(simulator.get_input_val(out), 0.into());
+
+        println!("<un_clock (already in reset)>");
+        simulator.un_clock();
+        println!("sim_state {:?}", simulator.sim_state);
+        assert_eq!(simulator.cycle, 1);
+        assert_eq!(simulator.get_input_val(out), 0.into());
+
+        println!("<clock>");
+        simulator.clock();
+        println!("sim_state {:?}", simulator.sim_state);
+        assert_eq!(simulator.cycle, 2);
+        assert_eq!(simulator.get_input_val(out), 1.into());
+
+        println!("<clock>");
+        simulator.clock();
+        println!("sim_state {:?}", simulator.sim_state);
+        assert_eq!(simulator.cycle, 3);
+        assert_eq!(simulator.get_input_val(out), 2.into());
+
+        println!("<reset>");
+        simulator.reset();
+        println!("sim_state {:?}", simulator.sim_state);
+        assert_eq!(simulator.cycle, 1);
+        assert_eq!(simulator.get_input_val(out), 0.into());
     }
 }